# New module @turf/linestring-to-polygon