--- a/packages/turf-linestring-to-polygon/bench.js
+++ b/packages/turf-linestring-to-polygon/bench.js
@@ -21,9 +21,10 @@ let fixtures = fs.readdirSync(directory).map(filename => {
  * geometry-linestring x 6,574,088 ops/sec ±3.62% (83 runs sampled)
  * linestring-incomplete x 6,768,527 ops/sec ±3.60% (84 runs sampled)
  * linestring x 6,752,969 ops/sec ±1.94% (84 runs sampled)
- * multi-linestring-incomplete x 3,263,278 ops/sec ±2.56% (88 runs sampled)
- * multi-linestring-with-hole x 3,381,728 ops/sec ±1.75% (88 runs sampled)
- * multi-linestrings-with-holes x 1,084,043 ops/sec ±2.28% (83 runs sampled)
+ * multi-linestring-incomplete x 808,779 ops/sec ±8.86% (80 runs sampled)
+ * multi-linestring-outer-ring-middle-position x 664,121 ops/sec ±1.52% (83 runs sampled)
+ * multi-linestring-with-hole x 1,018,657 ops/sec ±1.35% (86 runs sampled)
+ * multi-linestrings-with-holes x 421,758 ops/sec ±0.92% (88 runs sampled)
  */
 const suite = new Benchmark.Suite('turf-linestring-to-polygon');
 for (const {name, geojson} of fixtures) {