--- a/packages/turf-linestring-to-polygon/index.js
+++ b/packages/turf-linestring-to-polygon/index.js
@@ -1,15 +1,18 @@
+var bbox = require('@turf/bbox');
 var getCoords = require('@turf/invariant').getCoords;
 var helpers = require('@turf/helpers');
 var polygon = helpers.polygon;
 var multiPolygon = helpers.multiPolygon;
+var lineString = helpers.lineString;
 
 /**
  * Converts (Multi)LineString(s) to Polygon(s).
  *
  * @name lineStringToPolygon
  * @param {FeatureCollection|Feature<LineString|MultiLineString>} lines Features to convert
- * @param {boolean} [autoComplete=true] auto complete linestrings (matches first & last coordinates)
  * @param {Object} [properties] translates GeoJSON properties to Feature
+ * @param {boolean} [autoComplete=true] auto complete linestrings (matches first & last coordinates)
+ * @param {boolean} [orderCoords=true] sorts linestrings to place outer ring at the first position of the coordinates
  * @returns {Feature<Polygon|MultiPolygon>} converted to Polygons
  * @example
  * var line = {
@@ -25,12 +28,13 @@ var multiPolygon = helpers.multiPolygon;
  * //addToMap
  * var addToMap = [polygon];
  */
-module.exports = function (lines, autoComplete, properties) {
+module.exports = function (lines, properties, autoComplete, orderCoords) {
     // validation
     if (!lines) throw new Error('lines is required');
 
     // default params
     autoComplete = (autoComplete !== undefined) ? autoComplete : true;
+    orderCoords = (orderCoords !== undefined) ? orderCoords : true;
     var type = geomType(lines);
 
     switch (type) {
@@ -39,11 +43,11 @@ module.exports = function (lines, autoComplete, properties) {
         var coords = [];
         var features = (lines.features) ? lines.features : lines.geometries;
         features.forEach(function (line) {
-            coords.push(getCoords(lineStringToPolygon(line, autoComplete)));
+            coords.push(getCoords(lineStringToPolygon(line, {}, autoComplete, orderCoords)));
         });
         return multiPolygon(coords, properties);
     }
-    return lineStringToPolygon(lines, autoComplete, properties);
+    return lineStringToPolygon(lines, properties, autoComplete, orderCoords);
 };
 
 /**
@@ -51,11 +55,12 @@ module.exports = function (lines, autoComplete, properties) {
  *
  * @private
  * @param {Feature<LineString|MultiLineString>} line line
- * @param {boolean} [autoComplete=true] auto complete linestrings
  * @param {Object} [properties] translates GeoJSON properties to Feature
+ * @param {boolean} [autoComplete=true] auto complete linestrings
+ * @param {boolean} [orderCoords=true] sorts linestrings to place outer ring at the first position of the coordinates
  * @returns {Feature<Polygon>} line converted to Polygon
  */
-function lineStringToPolygon(line, autoComplete, properties) {
+function lineStringToPolygon(line, properties, autoComplete, orderCoords) {
     properties = properties || line.properties || {};
     var coords = getCoords(line);
     var type = geomType(line);
@@ -67,10 +72,24 @@ function lineStringToPolygon(line, autoComplete, properties) {
         if (autoComplete) coords = autoCompleteCoords(coords);
         return polygon([coords], properties);
     case 'MultiLineString':
-        if (autoComplete) coords.forEach(function (coord) {
-            coord = autoCompleteCoords(coord);
+        var multiCoords = [];
+        var largestArea = 0;
+
+        coords.forEach(function (coord) {
+            if (autoComplete) coord = autoCompleteCoords(coord);
+
+            // Largest LineString to be placed in the first position of the coordinates array
+            if (orderCoords) {
+                var area = calculateArea(bbox(lineString(coord)));
+                if (area > largestArea) {
+                    multiCoords.unshift(coord);
+                    largestArea = area;
+                } else multiCoords.push(coord);
+            } else {
+                multiCoords.push(coord);
+            }
         });
-        return polygon(coords, properties);
+        return polygon(multiCoords, properties);
     default:
         throw new Error('geometry type ' + type + ' is not supported');
     }
@@ -97,4 +116,19 @@ function autoCompleteCoords(coords) {
         coords.push(first);
     }
     return coords;
+}
+
+/**
+ * area - quick approximate area calculation (used to sort)
+ *
+ * @private
+ * @param {[number, number, number, number]} bbox BBox [west, south, east, north]
+ * @returns {number} very quick area calculation
+ */
+function calculateArea(bbox) {
+    var west = bbox[0];
+    var south = bbox[1];
+    var east = bbox[2];
+    var north = bbox[3];
+    return Math.abs(west - east) * Math.abs(south - north);
 }