--- a/packages/turf-linestring-to-polygon/package.json
+++ b/packages/turf-linestring-to-polygon/package.json
@@ -36,6 +36,7 @@
     "write-json-file": "^2.0.0"
   },
   "dependencies": {
+    "@turf/bbox": "^4.1.0",
     "@turf/helpers": "^4.1.0",
     "@turf/invariant": "^4.1.0"
   }