--- a/packages/turf-linestring-to-polygon/test.js
+++ b/packages/turf-linestring-to-polygon/test.js
@@ -22,7 +22,9 @@ let fixtures = fs.readdirSync(directories.in).map(filename => {
 
 test('turf-linestring-to-polygon', t => {
     for (const {name, filename, geojson} of fixtures) {
-        const results = lineStringToPolygon(geojson);
+        let {autoComplete, properties} = geojson.properties || {};
+        properties = properties || {stroke: '#F0F', 'stroke-width': 6};
+        const results = lineStringToPolygon(geojson, autoComplete, properties);
 
         if (process.env.REGEN) write.sync(directories.out + filename, results);
         t.deepEqual(load.sync(directories.out + filename), results, name);
@@ -30,6 +32,7 @@ test('turf-linestring-to-polygon', t => {
     // Handle Errors
     t.throws(() => lineStringToPolygon(point([10, 5])), 'throws - invalid geometry');
     t.throws(() => lineStringToPolygon(lineString([])), 'throws - empty coordinates');
+    t.throws(() => lineStringToPolygon(lineString([[10, 5], [20, 10], [30, 20]]), false), 'throws - autoComplete=false');
     t.end();
 });