--- a/packages/turf-linestring-to-polygon/test/in/collection-linestring.geojson
+++ b/packages/turf-linestring-to-polygon/test/in/collection-linestring.geojson
@@ -1,0 +1,69 @@
+{
+	"type": "FeatureCollection",
+	"features": [
+		{
+			"type": "Feature",
+			"properties": {
+				"name": "poly1",
+				"population": 400
+			},
+			"geometry": {
+				"type": "LineString",
+				"coordinates": [
+					[
+						102,
+						2
+					],
+					[
+						103,
+						2
+					],
+					[
+						103,
+						3
+					],
+					[
+						102,
+						3
+					],
+					[
+						102,
+						2
+					]
+				]
+			}
+		},
+		{
+			"type": "Feature",
+			"properties": {
+				"name": "poly1",
+				"population": 400
+			},
+			"geometry": {
+				"type": "LineString",
+				"coordinates": [
+					[
+						100,
+						0
+					],
+					[
+						101,
+						0
+					],
+					[
+						101,
+						1
+					],
+					[
+						100,
+						1
+					],
+					[
+						100,
+						0
+					]
+				]
+			}
+		}
+	]
+}