--- a/packages/turf-linestring-to-polygon/test/in/collection-linestring.geojson
+++ b/packages/turf-linestring-to-polygon/test/in/collection-linestring.geojson
@@ -4,8 +4,8 @@
 		{
 			"type": "Feature",
 			"properties": {
-				"name": "poly1",
-				"population": 400
+				"stroke": "#F00",
+				"stroke-width": 6
 			},
 			"geometry": {
 				"type": "LineString",
@@ -36,8 +36,8 @@
 		{
 			"type": "Feature",
 			"properties": {
-				"name": "poly1",
-				"population": 400
+				"stroke": "#00F",
+				"stroke-width": 6
 			},
 			"geometry": {
 				"type": "LineString",