--- a/packages/turf-linestring-to-polygon/test/in/linestring-incomplete.geojson
+++ b/packages/turf-linestring-to-polygon/test/in/linestring-incomplete.geojson
@@ -1,6 +1,9 @@
 {
   "type": "Feature",
-  "properties": {},
+  "properties": {
+    "stroke": "#F00",
+    "stroke-width": 6
+  },
   "geometry": {
     "type": "LineString",
     "coordinates": [