--- a/packages/turf-linestring-to-polygon/test/in/linestring.geojson
+++ b/packages/turf-linestring-to-polygon/test/in/linestring.geojson
@@ -2,8 +2,8 @@
 {
   "type": "Feature",
   "properties": {
-    "name": "poly1",
-    "population": 400
+    "stroke": "#F00",
+    "stroke-width": 6
   },
   "geometry": {
     "type": "LineString",