--- a/packages/turf-linestring-to-polygon/test/in/multi-linestring-with-hole.geojson
+++ b/packages/turf-linestring-to-polygon/test/in/multi-linestring-with-hole.geojson
@@ -1,8 +1,8 @@
 {
   "type": "Feature",
   "properties": {
-    "name": "poly1",
-    "population": 400
+    "stroke": "#F00",
+    "stroke-width": 6
   },
   "geometry": {
     "type": "MultiLineString",