--- a/packages/turf-linestring-to-polygon/test/in/multi-linestrings-with-holes.geojson
+++ b/packages/turf-linestring-to-polygon/test/in/multi-linestrings-with-holes.geojson
@@ -4,11 +4,8 @@
     {
       "type": "Feature",
       "properties": {
-        "stroke": "#0000ff",
-        "stroke-width": 6,
-        "stroke-opacity": 1,
-        "name": "poly1",
-        "population": 400
+        "stroke": "#00F",
+        "stroke-width": 6
       },
       "geometry": {
         "type": "MultiLineString",
@@ -63,11 +60,8 @@
     {
       "type": "Feature",
       "properties": {
-        "stroke": "#ff8000",
-        "stroke-width": 6,
-        "stroke-opacity": 1,
-        "name": "poly1",
-        "population": 400
+        "stroke": "#F80",
+        "stroke-width": 6
       },
       "geometry": {
         "type": "MultiLineString",
@@ -122,11 +116,8 @@
     {
       "type": "Feature",
       "properties": {
-        "stroke": "#008000",
-        "stroke-width": 6,
-        "stroke-opacity": 1,
-        "name": "poly1",
-        "population": 400
+        "stroke": "#080",
+        "stroke-width": 6
       },
       "geometry": {
         "type": "LineString",