--- a/packages/turf-linestring-to-polygon/test/out/geometry-linestring.geojson
+++ b/packages/turf-linestring-to-polygon/test/out/geometry-linestring.geojson
@@ -1,6 +1,9 @@
 {
 	"type": "Feature",
-	"properties": {},
+	"properties": {
+		"stroke": "#F0F",
+		"stroke-width": 6
+	},
 	"geometry": {
 		"type": "Polygon",
 		"coordinates": [