--- a/packages/turf-linestring-to-polygon/test/out/linestring-incomplete.geojson
+++ b/packages/turf-linestring-to-polygon/test/out/linestring-incomplete.geojson
@@ -1,6 +1,9 @@
 {
 	"type": "Feature",
-	"properties": {},
+	"properties": {
+		"stroke": "#F0F",
+		"stroke-width": 6
+	},
 	"geometry": {
 		"type": "Polygon",
 		"coordinates": [