--- a/packages/turf-linestring-to-polygon/test/out/linestring.geojson
+++ b/packages/turf-linestring-to-polygon/test/out/linestring.geojson
@@ -1,8 +1,8 @@
 {
 	"type": "Feature",
 	"properties": {
-		"name": "poly1",
-		"population": 400
+		"stroke": "#F0F",
+		"stroke-width": 6
 	},
 	"geometry": {
 		"type": "Polygon",