# New module 🎉  @turf/linestring-to-polygon

Converts (Multi)LineString(s) to Polygon(s).

> **Note**: Polygon with holes get converted to MultiLineString using `@turf/polygon-to-linestring`. This module is meant to work in harmony with `polygon-to-linestring`.

## To-Do

- [x] Feature => Polygon
- [x] FeatureCollection => MultiPolygon
- [x] Polygon outer ring must be in the first position of the Array

## JSDocs

```javascript
/**
 * Converts a {@link LineString} or {@link MultiLineString} to a {@link Polygon}.
 *
 * @name lineStringToPolygon
 * @param {FeatureCollection|Feature<LineString|MultiLineString>} lines Features to convert
 * @param {boolean} [autoComplete=true] auto complete linestrings
 * @returns {FeatureCollection|Feature<Polygon>} converted to Polygons
 * @example
 * var line = {
 *   'type': 'Feature',
 *   'properties': {},
 *   'geometry': {
 *     'type': 'LineString',
 *     'coordinates': [[125, -30], [145, -30], [145, -20], [125, -20], [125, -30]]
 *   }
 * }
 * var polygon = turf.lineStringToPolygon(line);
 *
 * //addToMap
 * var addToMap = [polygon];
 */
```
## Dependencies

```json
"dependencies": {
  "@turf/helpers": "^4.1.0",
  "@turf/invariant": "^4.1.0"
}
```

## Benchmark Results

```
collection-linestring x 2,337,816 ops/sec ±3.08% (86 runs sampled)
geometry-linestring x 6,574,088 ops/sec ±3.62% (83 runs sampled)
linestring-incomplete x 6,768,527 ops/sec ±3.60% (84 runs sampled)
linestring x 6,752,969 ops/sec ±1.94% (84 runs sampled)
multi-linestring-incomplete x 3,263,278 ops/sec ±2.56% (88 runs sampled)
multi-linestring-with-hole x 3,381,728 ops/sec ±1.75% (88 runs sampled)
multi-linestrings-with-holes x 1,084,043 ops/sec ±2.28% (83 runs sampled)
```

## Example 1 (with holes)

**Before:** Feature Collection of (Multi)LineString

![image](https://cloud.githubusercontent.com/assets/550895/25069130/30e521d6-2246-11e7-80b6-bbbd2fdcd035.png)

**After:** Feature Collection of Polygon (hole is preserved).

![image](https://cloud.githubusercontent.com/assets/550895/25069123/0cdfd006-2246-11e7-8338-f4ee8c13da41.png)

## Example 2 (incomplete lines)

**Before:** Incomplete (Multi)LineString

![image](https://cloud.githubusercontent.com/assets/550895/25069050/b2648dbc-2243-11e7-8146-58f7f9c30347.png)

**After:** Completed Polygons

![image](https://cloud.githubusercontent.com/assets/550895/25069134/635dea6c-2246-11e7-9df3-610eacc88e96.png)

CC: @Turfjs/ownership