@stebogit 2017-04-19
@DenisCarriere @rowanwins Man this is definitely gonna help for `isobands`! Yay!! 👍 
Just a note: in my experience and understanding of GeoJSON standards a [`multi-linestrings-with-holes`](https://github.com/Turfjs/turf/blob/7d27206c70ab1a241773781e85470b7e31fa02f0/packages/turf-linestring-to-polygon/test/out/multi-linestrings-with-holes.geojson) would be a `MultiPolygon`, not just a `Polygon`, and the outermost `Polygon` needs to be the first in the `coordinates` array. 
If I recall correctly [Google Maps](https://developers.google.com/maps/documentation/javascript/datalayer) gave me a hard time with just `Polygons` since it wasn't rendering the output correctly. [geojson.io](geojson.io) instead seems to be less strict, but still not always perfect output. 
What do you think, any experience that suggests differently?
I'd be happy to create some test so we can find out how it behaves.