@stebogit 2017-04-19
@DenisCarriere @rowanwins check this out: https://jsfiddle.net/jStefano/5xwxxco2/
if you render the same geojson in [geoson.io](geoson.io) instead it renders fine