@DenisCarriere 2017-04-19
> Man this is definitely gonna help for isobands 👍 !

😄 Indeed! This will help a lot of the modules that only handle LineStrings operations and need to easily convert them back into Polygons.

> in my experience and understanding of GeoJSON standards a multi-linestrings-with-holes would be a MultiPolygon

There's two very different tests going on here:
- [multi-linestrings-with-holes](https://github.com/Turfjs/turf/blob/linestring-to-polygon/packages/turf-linestring-to-polygon/test/out/multi-linestrings-with-hole.geojson) (Feature Collection of MultiLineStrings)
- [multi-linestring-with-holes](https://github.com/Turfjs/turf/blob/linestring-to-polygon/packages/turf-linestring-to-polygon/test/out/multi-linestring-with-hole.geojson) (Single Feature MultiLineString) 

## Current geometry transformations

- MultiLineString `=>` Polygon with Hole
- LineString `=>` Polygon
- FeatureCollection of MultiLineString `=>` FeatureCollection of Polygons with holes
- FeatureCollection of LineString `=>` FeatureCollection of Polygons

## Proposed geometry transformations (@stebogit's suggestion)

- MultiLineString `=>` Polygon with Hole
- LineString `=>` Polygon
- FeatureCollection of MultiLineString `=>` MultiPolygon with holes **(**changed)**
- FeatureCollection of LineString `=>` MultiPolygon **(**changed)**

I'm 👍 on this change, I prefer this proposed transformation.

## AutoComplete

> the outermost Polygon needs to be the first in the coordinates array.

Correct, however I've added a `autoComplete ` flag (default `true`) that completes the polygon if the first & last does not match (via [`autoCompleteCoords(coords)`](https://github.com/Turfjs/turf/blob/linestring-to-polygon/packages/turf-linestring-to-polygon/index.js#L80-L91)).

Sometimes a LineString will be "pretty close" but not exact, therefore it's pretty handy to have the `autoComplete` used by default.

CC: @stebogit