@DenisCarriere 2017-04-19
I've updated the repo with the proposed changes ( 👍 it's a lot better now).

## Key take aways

Output type changes based on input type

- Feature `=>` Polygon
- FeatureCollection `=>` MultiPolygon

The [`index.d.ts`](https://github.com/Turfjs/turf/blob/67071e21b10bc066b6b5fd1fd59d69f3b3f03382/packages/turf-linestring-to-polygon/index.d.ts) is a good place to understand the `input/output` flow.

@stebogit Feel free to include more test fixtures