@DenisCarriere 2017-04-19
> @stebogit check this out: https://jsfiddle.net/jStefano/5xwxxco2/
> if you render the same geojson in geojson.io instead it renders fine

Open an issue ticket to Google Maps API 🤓, that Polygon's geometry is perfectly valid.