@stebogit 2017-04-19
@DenisCarriere you know what, I did already at the time 🤓, and they told me the right geometry is a `MultiPolygon` (that's why I new about that, I'm not an expert! 😅)...
And actually it make sense, otherwise what's the need of an additional feature type? They would all be just `Polygons` no?
Is there any other geojson mapping tools we could use to see if that geometry is rendered as expected?