@DenisCarriere 2017-04-19
There's QGIS that you can open it with, just loaded your Polygon and it works just fine.

![image](https://cloud.githubusercontent.com/assets/550895/25198513/c40f5662-2515-11e7-99fc-f440815e0326.png)