@stebogit 2017-04-20
@DenisCarriere you are right, no need for `MultiPolygons` (although I wonder what's their use then...): a `Polygon` can contain multiple `LinearRings` coordinates, but as said the outermost ring MUST (according to the [specs](https://tools.ietf.org/html/rfc7946#section-3.1.6)) be in the first position of the coords array.
[Here](https://jsfiddle.net/jStefano/yd4cgy6b/) I modified that way the geojson passed to Google Maps and in fact it works fine also in there  🎉 
Google evidently is just more strict with geojson specs than others; however I believe it's the right way in order to avoid possible rendering issues.

My current approach (in `@turf/isobands`) to properly group the rings is first to order them by area; then, starting from the larger not grouped one, I collect all the rings (not yet grouped) inside that ring (using `@turf/inside` on each point of the ring in exam).
It is probably not the most efficient way, but it has been proved effective.

Any thought?

cc: @rowanwins