@DenisCarriere 2017-04-21
> a Polygon can contain multiple LinearRings coordinates, but as said the outermost ring MUST be in the first position of the coords array.

👍 Agreed, I'll add that in my next commit, a solution that won't be so computational heavy would be to retrieve all the `bbox` from each linestring ring (use [GeoJSON `bbox`](https://tools.ietf.org/html/rfc7946#section-5) property if exists). Then using only `bbox` as polygons we can quickly calculate the largest extent and define it as the outer ring (first position of array).

Just recently, I've improved `@turf/inside` with [`inBBox`](https://github.com/Turfjs/turf/blob/master/packages/turf-inside/index.js#L44) which quickly returns `false` if point feature is not inside bbox, which was inspired by [`rbush`](https://github.com/mourner/rbush). We can easily modify this function to reflect `inBBox(bbox, bbox) => true/false`.

## Example:

**Red Lines:** LineString input
**Blue Polys:** Polygon generated from `bbox` of input LineString

![image](https://cloud.githubusercontent.com/assets/550895/25283554/265b0434-2682-11e7-9d37-d73467f24c5a.png)

CC: @stebogit @rowanwins