@stebogit 2017-04-21
@DenisCarriere that is a brilliant idea! 👍 
However, what about this example (sorry, just to play devil's advocate):
![screen shot 2017-04-21 at 8 38 20 am](https://cloud.githubusercontent.com/assets/12717225/25285108/01004668-266e-11e7-8832-51ab806c3186.png)

cc: @rowanwins