@stebogit 2017-04-21
@DenisCarriere what if you simply "remove" the larger polygon (that you identify calculating the area) from the smaller one, you should be returned a null object, if the larger one includes the smaller right? If you are left with anything you know you have only a partial overlapping/inclusion.
I don't know though how `@turf/difference` would respond to this kind of clipping.

cc: @rowanwins